+++ packages/geoinfo/client/Geoinfo.js
@@ -2,43 +2,7 @@
 //move this module in client AND server
 
 Kepler.Geoinfo = {
-
-	loadByLoc: function(loc, cb) {
-
-/*		var showFields = {
-			"loc": true,
-			"near": true,
-			//"come": true,
-			"prov": true,
-			//"reg": true,
-			"naz": true
-		};*/
-
-
-		Meteor.call('findGeoinfoByLoc', loc, null, function(err, data) {
-
-			if(err){
-				console.log('findGeoinfoByLoc',err)
-			}
-			else if(data) {
-
-
-				var feature = K.Util.geo.createFeature('Point', [loc[1],loc[0]], data);
-				feature.templateMarker = 'markerGeoinfo';
-				feature.templatePopup = 'popupGeojson_geoinfo';
-
-				var geojson = K.Util.geo.createFeatureColl([feature]);
-
-				K.Map.hideCursor();
-				K.Map.addGeojson(geojson, {noFitBounds: true}, function() {
-					K.Map.layers.geojson.invoke('openPopup');
-				});
-			}
-			if(_.isFunction(cb))
-				cb(geojson);
-		});
-	},
-
+	
 	suncalc: function(loc) {
 		//https://github.com/mourner/suncalc/issues/101
 
@@ -59,20 +23,33 @@ Kepler.Geoinfo = {
 			sunset: format(times.sunset)
 		};
 	},
-	
-	geocoding: function(text, cb) {
-		
-		//var ret = new ReactiveVar([]);
 
-		Meteor.call('findGeocoding', text, function(err, data) {
-			if(err)
-	            console.log('Geoinfo',err);
-	        else
-	            cb(data);//ret.set(data);
+	loadByLoc: function(loc, cb) {
+
+		Meteor.call('findGeoinfoByLoc', loc, null, function(err, data) {
+
+			if(err){
+				console.log('findGeoinfoByLoc',err)
+			}
+			else if(data) {
+
+				data = _.extend(data, K.Geoinfo.suncalc(loc));
+
+				var feature = K.Util.geo.createFeature('Point', [loc[1],loc[0]], data);
+				feature.templateMarker = 'markerGeoinfo';
+				feature.templatePopup = 'popupGeojson_geoinfo';
+
+				var geojson = K.Util.geo.createFeatureColl([feature]);
+
+				K.Map.hideCursor();
+				K.Map.addGeojson(geojson, {noFitBounds: true}, function() {
+					K.Map.layers.geojson.invoke('openPopup');
+				});
+			}
+			if(_.isFunction(cb))
+				cb(geojson);
 		});
-
-		//return ret.get();
-	}
+	},
 	/*
 	//TODO
 	getAzimut: function(loc) {