+++ packages/edit/package.js
@@ -1,4 +1,4 @@
-var version = '1.6.0';
+var version = '1.6.1';
 
 Package.describe({
   version: version,